# Optimize delta decoding (part 5)