--- a/internal/bitpack/unpack_test.go
+++ b/internal/bitpack/unpack_test.go
@@ -39,7 +39,7 @@ func TestUnpackInt32(t *testing.T) {
 				bitpack.UnpackInt32(dst[:n], src, bitWidth)
 
 				if !reflect.DeepEqual(block[:n], dst[:n]) {
-					t.Errorf("values mismatch for length=%d\nwant: %v\ngot:  %v", n, block[:n], dst[:n])
+					t.Fatalf("values mismatch for length=%d\nwant: %v\ngot:  %v", n, block[:n], dst[:n])
 				}
 			}
 		})
@@ -72,7 +72,7 @@ func TestUnpackInt64(t *testing.T) {
 				bitpack.UnpackInt64(dst[:n], src, bitWidth)
 
 				if !reflect.DeepEqual(block[:n], dst[:n]) {
-					t.Errorf("values mismatch for length=%d\nwant: %v\ngot:  %v", n, block[:n], dst[:n])
+					t.Fatalf("values mismatch for length=%d\nwant: %v\ngot:  %v", n, block[:n], dst[:n])
 				}
 			}
 		})