This PR contributes to #226 by adding optimizations for decoding of 64 bit delta-encoded values.

I left comments inline, and focused on bit-packing between sizes 1 to 32 where the algorithm is a bit simpler than when values span across more than 32 bits. In practice, our use cases for delta-encoding on 64 bit values to have deltas <= MaxInt32 (e.g. encoding nearby timestamps which microsecond precision), we can revisit later and add specializations for the upper sizes if we measure it being a bottleneck.

I took a slightly different approach than in #230, I left documentation in the code explaining why.

```
name                     old time/op    new time/op     delta
UnpackInt64/bitWidth=1      163ns ± 0%       23ns ± 0%   -86.01%  (p=0.000 n=10+9)
UnpackInt64/bitWidth=8      164ns ± 0%       22ns ± 0%   -86.30%  (p=0.000 n=10+9)
UnpackInt64/bitWidth=10     186ns ± 0%       23ns ± 2%   -87.52%  (p=0.000 n=10+10)
UnpackInt64/bitWidth=20     204ns ± 1%       23ns ± 0%   -88.65%  (p=0.000 n=10+10)
UnpackInt64/bitWidth=30     241ns ± 1%       23ns ± 0%   -90.29%  (p=0.000 n=10+10)
UnpackInt64/bitWidth=40     252ns ± 1%      255ns ± 0%    +1.46%  (p=0.000 n=10+6)

name                     old speed      new speed       delta
UnpackInt64/bitWidth=1   3.14GB/s ± 0%  22.42GB/s ± 0%  +614.85%  (p=0.000 n=10+9)
UnpackInt64/bitWidth=8   3.13GB/s ± 0%  22.85GB/s ± 0%  +629.74%  (p=0.000 n=10+9)
UnpackInt64/bitWidth=10  2.76GB/s ± 0%  22.10GB/s ± 2%  +700.99%  (p=0.000 n=10+10)
UnpackInt64/bitWidth=20  2.51GB/s ± 1%  22.15GB/s ± 0%  +781.28%  (p=0.000 n=10+10)
UnpackInt64/bitWidth=30  2.13GB/s ± 1%  21.92GB/s ± 0%  +930.28%  (p=0.000 n=10+10)
UnpackInt64/bitWidth=40  2.03GB/s ± 1%   2.01GB/s ± 0%    -1.45%  (p=0.000 n=10+6)
```

```
name                                  old time/op  new time/op  delta
MergeFiles/INT64/groups=2,rows=20000  9.94µs ± 1%  7.48µs ± 1%  -24.76%  (p=0.000 n=9+9)

name                                  old row/s    new row/s    delta
MergeFiles/INT64/groups=2,rows=20000   97.4M ± 1%  129.4M ± 1%  +32.92%  (p=0.000 n=9+9)
```